style(list): fix code review comment
fix #463
